Simplfy ASIOInterface
