+++ bcos-gateway/bcos-gateway/libnetwork/ASIOInterface.h
@@ -63,30 +63,57 @@ public:
     // bridge frame, and no per-operation allocation — and are obtained from the non-virtual
     // awaitable* helpers below.
     //
-    // The read mock point is the injectable read-initiation hook (setReadSomeInitiate):
-    // production uses the default target (the real async_read_some dispatch, see the private
-    // initiateReadSome); read-loop test fakes replace it to park / control read completions
-    // deterministically. CONTRACT (for injected fakes and for every initiate call in this
-    // header): the completion must be handed to a deferred executor (asio, or a post to some
-    // io_context) — it must be neither invoked nor dropped synchronously. A synchronous
-    // invocation / drop is neutralized by the arm/cancel handshake in AsioAwaitable (see
-    // AsioAwaitable.h) rather than corrupting the running coroutine, but the awaitable's
+    // The read path is a COMPILE-TIME policy (the template parameter of awaitableReadSome):
+    // production uses DefaultReadPolicy, whose invoke() directly dispatches async_read_some on
+    // the socket (TCP vs SSL, see DefaultReadPolicy below) — no std::function, no virtual call,
+    // fully inlined. Read-loop test fakes substitute their own policy type to park / control
+    // read completions deterministically. CONTRACT (for custom policies and for every initiate
+    // call in this header): the completion must be handed to a deferred executor (asio, or a
+    // post to some io_context) — it must be neither invoked nor dropped synchronously. A
+    // synchronous invocation / drop is neutralized by the arm/cancel handshake in AsioAwaitable
+    // (see AsioAwaitable.h) rather than corrupting the running coroutine, but the awaitable's
     // total-completion guarantee is clearest when every initiation defers.
     using ReadSomeHandler = detail::AsioCompletion<boost::system::error_code, std::size_t>;
-    using ReadSomeInitiate = std::function<void(const std::shared_ptr<SocketFace>& socket,
-        boost::asio::mutable_buffer buffers, ReadSomeHandler completion)>;
-    // unit-test seam for the read path (see the contract above); set before reads are armed
-    void setReadSomeInitiate(ReadSomeInitiate initiate)
-    {
-        m_readSomeInitiate = std::move(initiate);
-    }
 
+    // Production read-initiation policy: directly dispatches async_read_some on the socket
+    // (TCP vs SSL, with the unexpected-type default completing via operation_not_supported).
+    // Nested in ASIOInterface so it can read the private m_type; tests provide their own policy
+    // with the same invoke() signature.
+    struct DefaultReadPolicy
+    {
+        static void invoke(ASIOInterface* self, const std::shared_ptr<SocketFace>& socket,
+            boost::asio::mutable_buffer buffers, ReadSomeHandler completion)
+        {
+            switch (self->m_type)
+            {
+            case TCP_ONLY:
+                socket->ref().async_read_some(buffers, std::move(completion));
+                break;
+            case SSL:
+                socket->sslref().async_read_some(buffers, std::move(completion));
+                break;
+            default:
+                // total completion: an unexpected type must still answer the read, or the
+                // awaiting read-loop coroutine pins forever. POST the completion rather than
+                // invoking it inline — this runs on the initiator's stack inside await_suspend,
+                // and an inline invocation would resume the coroutine from within its own
+                // await_suspend.
+                boost::asio::post(socket->ioService(),
+                    [completion = std::move(completion)]() mutable {
+                        completion(boost::asio::error::operation_not_supported, std::size_t{0});
+                    });
+                break;
+            }
+        }
+    };
+
+    template <typename ReadPolicy = DefaultReadPolicy>
     auto awaitableReadSome(
         const std::shared_ptr<SocketFace>& socket, boost::asio::mutable_buffer buffers)
     {
         return makeAsioAwaitable<boost::system::error_code, std::size_t>(
             [this, socket, buffers](auto handler) {
-                m_readSomeInitiate(socket, buffers, std::move(handler));
+                ReadPolicy::invoke(this, socket, buffers, std::move(handler));
             });
     }
 
@@ -173,12 +200,6 @@ public:
     }
 
 private:
-    // Real read initiation: async_read_some dispatch on m_type (TCP vs SSL, with the unexpected-
-    // type default completing via operation_not_supported). This is the default target of
-    // m_readSomeInitiate (see the seam contract on setReadSomeInitiate).
-    void initiateReadSome(const std::shared_ptr<SocketFace>& socket,
-        boost::asio::mutable_buffer buffers, ReadSomeHandler completion);
-
     // resolve + connect helper backing awaitableResolveConnect (total completion: the handler
     // fires with an error when resolution fails — see the coroutine-interface comment above).
     // Templated because the handler is the move-only AsioCompletion.
@@ -217,8 +238,5 @@ private:
     std::optional<ba::ssl::context> m_srvContext;
     std::optional<ba::ssl::context> m_clientContext;
     int m_type = 0;
-    // The read-initiation seam (see setReadSomeInitiate): defaults to initiateReadSome; read
-    // fakes replace it before reads are armed.
-    ReadSomeInitiate m_readSomeInitiate;
 };
 }  // namespace bcos::gateway
